refactor: Extract GameElement into @bruff/game-element package
Moves the GameElement web-component base class and its tests out of
@bruff/game/lib into a new source-only workspace package, mirroring the
@bruff/utils conventions (index.ts barrel, module/ layout, browser vitest
with 100% coverage thresholds). @bruff/game now consumes it via the
workspace symlink so the extraction is transparent to the built artifact.
