+++ packages/game-element/.prettierignore
@@ -1,0 +1,3 @@
+coverage
+test-results
+playwright-report
