+++ packages/game-element/eslint.config.js
@@ -1,0 +1,8 @@
+import bruffEslintConfig from "@bruff/eslint-config";
+
+export default [
+  {
+    ignores: ["**/*.css", "**/*.scss", "coverage/**/*.*"],
+  },
+  ...bruffEslintConfig,
+];
