+++ packages/game-element/index.ts
@@ -1,0 +1,1 @@
+export { GameElement } from "./module/game-element.js";
