+++ packages/game-element/tsconfig.json
@@ -1,0 +1,12 @@
+{
+  "extends": "../../tsconfig.base.json",
+  "compilerOptions": {
+    "baseUrl": ".",
+    "composite": true,
+    "paths": {
+      "@/*": ["module/*"]
+    }
+  },
+  "include": ["index.ts", "module/**/**.ts", "vitest.config.ts"],
+  "exclude": ["node_modules", "eslint.config.js"]
+}
